feat(#2642): eo syntax
