+++ eo-maven-plugin/src/main/resources/log4j.properties
@@ -6,7 +6,7 @@ log4j.appender.CONSOLE.layout.ConversionPattern=%d{HH:mm:ss} [%color{%p}] %c: %m
 
 log4j.logger.org.eolang=INFO
 log4j.logger.com.yegor256.xsline=INFO
-log4j.logger.org.eolang.parser.Syntax=INFO
+log4j.logger.org.eolang.parser.EoSyntax=INFO
 log4j.logger.org.eolang.parser.Scenario=INFO
 log4j.logger.org.eolang.parser.XMIRTest=INFO
 log4j.logger.org.eolang.maven.SodgMojo=TRACE
