+++ eo-maven-plugin/src/test/resources/log4j.properties
@@ -6,7 +6,7 @@ log4j.appender.CONSOLE.layout.ConversionPattern=%d{HH:mm:ss} [%p] %c: %m%n
 
 log4j.logger.org.eolang=INFO
 log4j.logger.com.jcabi.log.VerboseProcess=INFO
-log4j.logger.org.eolang.parser.Syntax=INFO
+log4j.logger.org.eolang.parser.EoSyntax=INFO
 log4j.logger.org.eolang.parser.Program=INFO
 log4j.logger.org.eolang.maven.SodgMojo=INFO
 log4j.logger.org.eolang.maven.SodgMojoTest=INFO
