+++ eo-parser/src/main/java/org/eolang/parser/PhiSyntax.java
@@ -1,0 +1,15 @@
+package org.eolang.parser;
+
+import com.jcabi.xml.XML;
+import java.io.IOException;
+
+/**
+ * Syntax parser, from Phi-calculus to XMIR, using ANTLR4.
+ * @since 0.34.0
+ */
+public class PhiSyntax implements Syntax {
+    @Override
+    public XML parsed() throws IOException {
+        return null;
+    }
+}
