+++ eo-runtime/src/main/java/EOorg/EOeolang/EOio/EOstdout.java
@@ -27,8 +27,6 @@
  */
 package EOorg.EOeolang.EOio;
 
-import EOorg.EOeolang.EObool;
-import EOorg.EOeolang.EObytes;
 import java.io.PrintStream;
 import org.eolang.AtFree;
 import org.eolang.AtLambda;
@@ -36,7 +34,6 @@ import org.eolang.Attr;
 import org.eolang.Data;
 import org.eolang.Param;
 import org.eolang.PhDefault;
-import org.eolang.PhWith;
 import org.eolang.Phi;
 import org.eolang.Versionized;
 import org.eolang.XmirObject;
@@ -76,7 +73,9 @@ public class EOstdout extends PhDefault {
             new AtLambda(
                 this,
                 rho -> {
-                    out.print(new Param(rho, "text").strong(String.class));
+                    out.print(
+                        new Param(rho, "text").strong(String.class)
+                    );
                     return new Data.ToPhi(true);
                 }
             )
